backend/requirements.txt
# FastAPI + servidor
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Base de datos
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Validación y settings
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utilidades
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
cloudinary==1.41.0
